Simplify InputBatchingDecorator to be one per vajram and handle epochs internally rather than creating one decorator per epoch per vajram
