+++ krystex/src/main/java/com/flipkart/krystal/krystex/batching/BatchSizeSupplier.java
@@ -1,0 +1,8 @@
+package com.flipkart.krystal.krystex.batching;
+
+import com.flipkart.krystal.core.VajramID;
+
+@FunctionalInterface
+public interface BatchSizeSupplier {
+  int getBatchSize(VajramID vajramId);
+}
