+++ krystex/src/main/java/com/flipkart/krystal/krystex/batching/InputBatcherConfig.java
@@ -1,12 +1,298 @@
 package com.flipkart.krystal.krystex.batching;
 
+import static com.flipkart.krystal.krystex.batching.DepChainBatcherConfig.sharedBatcher;
+
 import com.flipkart.krystal.core.VajramID;
+import com.flipkart.krystal.facets.Dependency;
+import com.flipkart.krystal.facets.Facet;
+import com.flipkart.krystal.krystex.DependentChainDisabler;
+import com.flipkart.krystal.krystex.VajramGraph;
+import com.flipkart.krystal.krystex.batching.LogicSet.DepResolvers;
+import com.flipkart.krystal.krystex.batching.LogicSet.OutputLogics;
+import com.flipkart.krystal.krystex.kryon.DependentChain;
+import com.flipkart.krystal.krystex.logicdecoration.OutputLogicDecorator;
+import com.flipkart.krystal.traits.TraitDispatchPolicies;
+import com.flipkart.krystal.traits.TraitDispatchPolicy;
+import com.flipkart.krystal.vajram.IOVajramDef;
+import com.flipkart.krystal.vajram.batching.InputBatcher;
+import com.flipkart.krystal.vajram.batching.InputBatcherImpl;
+import com.flipkart.krystal.vajram.exec.VajramDefinition;
+import com.flipkart.krystal.vajram.facets.specs.FacetSpec;
 import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableMap;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.LinkedHashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Map.Entry;
+import java.util.Set;
+import java.util.function.Function;
 import lombok.Builder;
-import lombok.Singular;
+import org.checkerframework.checker.nullness.qual.NonNull;
+import org.jspecify.annotations.Nullable;
 
 @Builder
 public record InputBatcherConfig(
-    @Singular
-        ImmutableMap<VajramID, ImmutableList<DepChainBatcherConfig>> depChainBatcherConfigs) {}
+    Function<VajramID, @Nullable OutputLogicDecorator> decoratorFactory) {
+
+  public static InputBatcherConfig computeSharedBatcherConfig(
+      VajramGraph graph,
+      BatchSizeSupplier batchSizeSupplier,
+      TraitDispatchPolicies traitDispatchPolicies,
+      DependentChainDisabler dependentChainDisabler,
+      Collection<VajramID> externallyInvocableVajramIds) {
+    List<VajramDefinition> externallyInvocableVajrams =
+        externallyInvocableVajramIds.stream().map(graph::getVajramDefinition).toList();
+    Map<VajramID, Map<Integer, Set<DependentChain>>> ioVajramsToOrdinalChains = new HashMap<>();
+    Map<VajramID, Integer> vajramsToOutgoingOrdinals = new HashMap<>();
+    for (VajramDefinition vajramDefinition : externallyInvocableVajrams) {
+      Collection<VajramID> dispatchTargets =
+          getDispatchTargets(vajramDefinition.vajramId(), graph, traitDispatchPolicies);
+      for (VajramID dispatchTargetID : dispatchTargets) {
+        collateDepChainOrdinals(
+            ioVajramsToOrdinalChains,
+            vajramsToOutgoingOrdinals,
+            graph,
+            dispatchTargetID,
+            graph.kryonDefinitionRegistry().getDependentChainsStart(),
+            0,
+            dependentChainDisabler,
+            traitDispatchPolicies);
+      }
+    }
+    return registerBatchers(ioVajramsToOrdinalChains, batchSizeSupplier, graph);
+  }
+
+  static Collection<VajramID> getDispatchTargets(
+      VajramID depVajramID, VajramGraph graph, TraitDispatchPolicies traitDispatchPolicies) {
+    VajramDefinition depVajramDef = graph.getVajramDefinition(depVajramID);
+    Collection<VajramID> depVajramIDs = new ArrayList<>();
+    if (depVajramDef.isTrait()) {
+      TraitDispatchPolicy traitDispatchPolicy = traitDispatchPolicies.get(depVajramID);
+      if (traitDispatchPolicy == null) {
+        throw new IllegalStateException(
+            "Trait "
+                + depVajramID
+                + " does not have a trait dispatch policy defined. Cannot auto-compute batcher config.");
+      }
+      for (VajramID vajramID : traitDispatchPolicy.dispatchTargetIDs()) {
+        depVajramIDs.addAll(getDispatchTargets(vajramID, graph, traitDispatchPolicies));
+      }
+    } else {
+      depVajramIDs = List.of(depVajramID);
+    }
+    return depVajramIDs;
+  }
+
+  /**
+   * Given A LogicSet of a given vajram (ex: resolvers of a dependency or output logics) and an
+   * incoming ordinal, this method computes the effective ordinal of all the facets which are the
+   * sources to the logic set.
+   *
+   * <p>For example: if a vajram has an output logic which has two sources which are dependency
+   * vajrams with response ordinals 1 and 2 respectively, then the source ordinal of the output
+   * logic will be
+   *
+   * <ul>
+   *   <li>2 if the two dependencies are parallel (max 1,2)
+   *   <li>3 if the two dependencies are sequential (1 + 2)
+   * </ul>
+   *
+   * @param sourceOrdinalKey
+   * @param sourceOrdinals
+   * @param vajramsToResponseOrdinals
+   * @param graph
+   * @param traitDispatchPolicies
+   * @return
+   */
+  static int computeSourceOrdinal(
+      SourceOrdinalKey sourceOrdinalKey,
+      Map<SourceOrdinalKey, Integer> sourceOrdinals,
+      Map<VajramID, Integer> vajramsToResponseOrdinals,
+      VajramGraph graph,
+      TraitDispatchPolicies traitDispatchPolicies) {
+    if (sourceOrdinals.containsKey(sourceOrdinalKey)) {
+      return sourceOrdinals.get(sourceOrdinalKey);
+    }
+    LogicSet logicSet = sourceOrdinalKey.logicSet();
+    int incomingOrdinal = sourceOrdinalKey.incomingOrdinal();
+    int sourceOrdinal = incomingOrdinal;
+    for (Facet source : logicSet.sources(graph)) {
+      if (source instanceof Dependency sourceDependency) {
+        int i =
+            computeSourceOrdinal(
+                new SourceOrdinalKey(new DepResolvers(sourceDependency), incomingOrdinal),
+                sourceOrdinals,
+                vajramsToResponseOrdinals,
+                graph,
+                traitDispatchPolicies);
+        for (VajramID sourceDispatchTarget :
+            getDispatchTargets(sourceDependency.onVajramID(), graph, traitDispatchPolicies)) {
+          sourceOrdinal =
+              Math.max(
+                  sourceOrdinal,
+                  i
+                      + computeResponseOrdinal(
+                          sourceDispatchTarget,
+                          sourceOrdinals,
+                          vajramsToResponseOrdinals,
+                          graph,
+                          traitDispatchPolicies));
+        }
+      }
+    }
+    sourceOrdinals.put(sourceOrdinalKey, sourceOrdinal);
+    return sourceOrdinal;
+  }
+
+  /**
+   * Given a vajram, computes the response ordinal of that vajram in isolation (irrespective of
+   * where it's invoked from - i.e. assuming incoming ordinal is 0)
+   *
+   * @param vajramBeingInvokedID
+   * @param sourceOrdinals
+   * @param vajramsToResponseOrdinals
+   * @param graph
+   * @param traitDispatchPolicies
+   * @return
+   */
+  static int computeResponseOrdinal(
+      VajramID vajramBeingInvokedID,
+      Map<SourceOrdinalKey, Integer> sourceOrdinals,
+      Map<VajramID, Integer> vajramsToResponseOrdinals,
+      VajramGraph graph,
+      TraitDispatchPolicies traitDispatchPolicies) {
+    if (!vajramsToResponseOrdinals.containsKey(vajramBeingInvokedID)) {
+      int responseOrdinal =
+          computeSourceOrdinal(
+              new SourceOrdinalKey(new OutputLogics(vajramBeingInvokedID), 0),
+              sourceOrdinals,
+              vajramsToResponseOrdinals,
+              graph,
+              traitDispatchPolicies);
+      if (graph.getVajramDefinition(vajramBeingInvokedID).def() instanceof IOVajramDef<?>) {
+        responseOrdinal++;
+      }
+      vajramsToResponseOrdinals.put(vajramBeingInvokedID, responseOrdinal);
+    }
+    return vajramsToResponseOrdinals.get(vajramBeingInvokedID);
+  }
+
+  /**
+   * For every IO Vajram, this method collates all depChains ending in that IO Vajram by its ordinal
+   *
+   * @param ioVajramsToOrdinalChains a cache which maps a vajram to depth(ordinal) to set of
+   *     depChains which map to that depth
+   * @param vajramsToResponseOrdinals a cache which maps a vajram to its response ordinals
+   * @param graph
+   * @param vajramIDBeingInvoked the vajram for which depchain ordinals need to be collated
+   * @param incomingDepChain
+   * @param incomingOrdinal the current ordinal at the invocation location from where this vajram is
+   *     being invoked
+   * @param dependentChainDisabler
+   * @param traitDispatchPolicies
+   */
+  static void collateDepChainOrdinals(
+      Map<VajramID, Map<Integer, Set<DependentChain>>> ioVajramsToOrdinalChains,
+      Map<VajramID, Integer> vajramsToResponseOrdinals,
+      VajramGraph graph,
+      VajramID vajramIDBeingInvoked,
+      DependentChain incomingDepChain,
+      int incomingOrdinal,
+      DependentChainDisabler dependentChainDisabler,
+      TraitDispatchPolicies traitDispatchPolicies) {
+    if (dependentChainDisabler.isDisabled(incomingDepChain)) {
+      return;
+    }
+    VajramDefinition vajramBeingInvoked = graph.getVajramDefinition(vajramIDBeingInvoked);
+    if (vajramBeingInvoked.isTrait()) {
+      throw new IllegalArgumentException(
+          "collateDepChainOrdinals cannot be called for traits. First resolve dispatch targets before calling this method.");
+    }
+    List<Dependency> dependencies = getDependencies(vajramBeingInvoked);
+    Map<SourceOrdinalKey, Integer> sourceOrdinals = new HashMap<>();
+    for (Dependency dependency : dependencies) {
+      DependentChain outgoingDepChain =
+          incomingDepChain.extend(vajramBeingInvoked.vajramId(), dependency);
+      for (VajramID depVajramID :
+          getDispatchTargets(dependency.onVajramID(), graph, traitDispatchPolicies)) {
+        collateDepChainOrdinals(
+            ioVajramsToOrdinalChains,
+            vajramsToResponseOrdinals,
+            graph,
+            depVajramID,
+            outgoingDepChain,
+            computeSourceOrdinal(
+                new SourceOrdinalKey(new DepResolvers(dependency), incomingOrdinal),
+                sourceOrdinals,
+                vajramsToResponseOrdinals,
+                graph,
+                traitDispatchPolicies),
+            dependentChainDisabler,
+            traitDispatchPolicies);
+      }
+    }
+
+    if (vajramBeingInvoked.def() instanceof IOVajramDef<?>) {
+      final int depChainOrdinal =
+          incomingOrdinal
+              + computeResponseOrdinal(
+                  vajramIDBeingInvoked,
+                  sourceOrdinals,
+                  vajramsToResponseOrdinals,
+                  graph,
+                  traitDispatchPolicies);
+      ioVajramsToOrdinalChains
+          .computeIfAbsent(vajramIDBeingInvoked, _vid -> new HashMap<>())
+          .computeIfAbsent(depChainOrdinal, _depth -> new HashSet<>())
+          .add(incomingDepChain);
+    }
+  }
+
+  static @NonNull List<Dependency> getDependencies(VajramDefinition vajramBeingInvoked) {
+    List<Dependency> dependencies =
+        vajramBeingInvoked.facetSpecs().stream()
+            .filter(f -> f instanceof Dependency)
+            .<@NonNull Dependency>map(Dependency.class::cast)
+            .toList();
+    return dependencies;
+  }
+
+  static InputBatcherConfig registerBatchers(
+      Map<VajramID, Map<Integer, Set<DependentChain>>> ioNodes,
+      BatchSizeSupplier batchSizeSupplier,
+      VajramGraph graph) {
+    Map<VajramID, ImmutableList<DepChainBatcherConfig>> depChainBatcherConfigs =
+        new LinkedHashMap<>();
+    ioNodes.forEach(
+        (vajramId, ioNodeMap) -> {
+          if (isBatchingNeededForIoVajram(graph, vajramId)) {
+            List<DepChainBatcherConfig> inputModulatorConfigs = new ArrayList<>(ioNodeMap.size());
+            for (Entry<Integer, Set<DependentChain>> entry : ioNodeMap.entrySet()) {
+              int depth = entry.getKey();
+              Set<DependentChain> depChains = entry.getValue();
+              inputModulatorConfigs.add(
+                  sharedBatcher(
+                      vajramId.id() + ":depth(" + depth + ")",
+                      depChains.toArray(DependentChain[]::new)));
+            }
+            depChainBatcherConfigs.put(vajramId, ImmutableList.copyOf(inputModulatorConfigs));
+          }
+        });
+    Function<VajramID, InputBatcher> inputBatcherFactory =
+        vajramID -> new InputBatcherImpl(batchSizeSupplier.getBatchSize(vajramID));
+    return new InputBatcherConfig(
+        vajramID ->
+            new InputBatchingDecorator(
+                () -> inputBatcherFactory.apply(vajramID), depChainBatcherConfigs.get(vajramID)));
+  }
+
+  static boolean isBatchingNeededForIoVajram(VajramGraph graph, VajramID vajramID) {
+    VajramDefinition vajramDefinition = graph.vajramDefinitions().get(vajramID);
+    return vajramDefinition != null && vajramDefinition.metadata().isBatched();
+  }
+
+  record SourceOrdinalKey(LogicSet logicSet, int incomingOrdinal) {}
+}
