+++ krystex/src/main/java/com/flipkart/krystal/krystex/batching/InputBatcherStrategy.java
@@ -1,13 +1,12 @@
 package com.flipkart.krystal.krystex.batching;
 
 import com.flipkart.krystal.krystex.KrystexGraph;
-import com.flipkart.krystal.krystex.batching.DepChainBatcherConfig.BatchSizeSupplier;
 
 public sealed interface InputBatcherStrategy {
 
   /**
    * This indicates that the {@link KrystexGraph} should use the default shared batcher config
-   * computated by {@link DepChainBatcherConfig}.
+   * computed by {@link DepChainBatcherConfig}.
    *
    * @param batchSizeSupplier the batch size supplier to use to determine batch size for vajrams
    */
