+++ krystex/src/main/java/com/flipkart/krystal/krystex/batching/InputBatchingDecorator.java
@@ -1,10 +1,7 @@
 package com.flipkart.krystal.krystex.batching;
 
 import static com.flipkart.krystal.except.KrystalCompletionException.wrapAsCompletionException;
-import static java.util.Collections.unmodifiableSet;
 
-import com.flipkart.krystal.config.ConfigProvider;
-import com.flipkart.krystal.config.NestedConfig;
 import com.flipkart.krystal.core.OutputLogicExecutionInput;
 import com.flipkart.krystal.data.ExecutionItem;
 import com.flipkart.krystal.krystex.OutputLogic;
@@ -14,16 +11,19 @@ import com.flipkart.krystal.krystex.decoration.FlushableDecorator;
 import com.flipkart.krystal.krystex.decoration.InitiableWithActiveDepChains;
 import com.flipkart.krystal.krystex.decoration.InitiateActiveDepChains;
 import com.flipkart.krystal.krystex.kryon.DependentChain;
+import com.flipkart.krystal.krystex.logicdecoration.LogicExecutionContext;
 import com.flipkart.krystal.krystex.logicdecoration.OutputLogicDecorator;
 import com.flipkart.krystal.vajram.batching.BatchEnabledFacetValues;
 import com.flipkart.krystal.vajram.batching.BatchedFacets;
 import com.flipkart.krystal.vajram.batching.InputBatcher;
-import com.google.common.collect.ImmutableSet;
+import com.google.common.collect.Sets;
 import java.util.ArrayList;
+import java.util.LinkedHashMap;
 import java.util.LinkedHashSet;
 import java.util.List;
+import java.util.Map;
 import java.util.Set;
-import java.util.function.Predicate;
+import java.util.function.Supplier;
 import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
 
 public final class InputBatchingDecorator
@@ -31,25 +31,36 @@ public final class InputBatchingDecorator
 
   public static final String DECORATOR_TYPE = InputBatchingDecorator.class.getName();
 
-  private final String instanceId;
-  private final InputBatcher inputBatcher;
-  private final Predicate<DependentChain> isApplicableToDependentChain;
-  private final Set<DependentChain> dependantChainsToFlush = new LinkedHashSet<>();
-  private @MonotonicNonNull Set<DependentChain> activeDependantChains;
+  private final List<InputBatcher> sharedInputBatchersByEpoch;
+  private final Map<DependentChain, InputBatcher> simpleInputBatchersByDepChain =
+      new LinkedHashMap<>();
+  private final Supplier<InputBatcher> inputBatcherFactory;
+  private final List<DepChainBatcherConfig> depChainBatcherConfigs;
+  private final Map<DependentChain, Integer> epochByDepChain = new LinkedHashMap<>();
+  private final List<Set<DependentChain>> dependentChainsToFlushByEpoch = new ArrayList<>();
   private @MonotonicNonNull OutputLogicExecutionInput outputLogicExecutionInput;
 
   public InputBatchingDecorator(
-      String instanceId,
-      InputBatcher inputBatcher,
-      Predicate<DependentChain> isApplicableToDependentChain) {
-    this.instanceId = instanceId;
-    this.inputBatcher = inputBatcher;
-    this.isApplicableToDependentChain = isApplicableToDependentChain;
+      Supplier<InputBatcher> inputBatcherFactory,
+      List<DepChainBatcherConfig> depChainBatcherConfigs) {
+    this.inputBatcherFactory = inputBatcherFactory;
+    this.sharedInputBatchersByEpoch = new ArrayList<>(depChainBatcherConfigs.size());
+    this.depChainBatcherConfigs = depChainBatcherConfigs;
+    for (int epoch = 0; epoch < depChainBatcherConfigs.size(); epoch++) {
+      this.sharedInputBatchersByEpoch.add(inputBatcherFactory.get());
+      for (DependentChain dependentChain : depChainBatcherConfigs.get(epoch).dependentChains()) {
+        epochByDepChain.put(dependentChain, epoch);
+      }
+    }
   }
 
   @Override
   public OutputLogic<Object> decorateLogic(
-      OutputLogic<Object> logicToDecorate, OutputLogicDefinition<Object> originalLogicDefinition) {
+      OutputLogic<Object> logicToDecorate,
+      OutputLogicDefinition<Object> originalLogicDefinition,
+      LogicExecutionContext context) {
+    DependentChain dependentChain = context.dependentChain();
+    InputBatcher inputBatcher = getInputBatcher(dependentChain);
     inputBatcher.onBatching(
         requests -> requests.forEach(request -> batchFacetsList(logicToDecorate, request)));
     return input -> {
@@ -62,8 +73,8 @@ public final class InputBatchingDecorator
               f -> {
                 if (!(f.facetValues() instanceof BatchEnabledFacetValues)) {
                   throw new IllegalStateException(
-                      "Expected to receive instance of BatchEnabledFacetValues in batcher %s but received %s"
-                          .formatted(instanceId, f));
+                      "Expected to receive instance of BatchEnabledFacetValues in batcher for %s but received %s"
+                          .formatted(context.vajramID(), f));
                 }
               });
       List<BatchedFacets> batchedFacetsList = new ArrayList<>();
@@ -76,33 +87,39 @@ public final class InputBatchingDecorator
     };
   }
 
+  private InputBatcher getInputBatcher(DependentChain dependentChain) {
+    Integer epoch = epochByDepChain.get(dependentChain);
+    if (epoch == null) {
+      return simpleInputBatchersByDepChain.computeIfAbsent(
+          dependentChain, _d -> inputBatcherFactory.get());
+    } else {
+      return sharedInputBatchersByEpoch.get(epoch);
+    }
+  }
+
   @Override
   public void flushDecorator(FlushCommand flushCommand) {
-    dependantChainsToFlush.remove(flushCommand.dependantsChain());
-    if (dependantChainsToFlush.isEmpty()) {
-      inputBatcher.batch();
-      dependantChainsToFlush.addAll(activeDependantChains());
+    DependentChain dependentChain = flushCommand.dependentChain();
+    Integer epoch = epochByDepChain.get(dependentChain);
+    Set<DependentChain> dependentChainsToFlush = Set.of();
+    if (epoch != null) {
+      dependentChainsToFlush = dependentChainsToFlushByEpoch.get(epoch);
+      dependentChainsToFlush.remove(dependentChain);
+    }
+    if (dependentChainsToFlush.isEmpty()) {
+      getInputBatcher(dependentChain).batch();
     }
   }
 
   @Override
   public void initiateActiveDepChains(InitiateActiveDepChains initiateActiveDepChains) {
-    if (activeDependantChains == null) {
-      Set<DependentChain> allActiveDepChains = initiateActiveDepChains.dependentChains();
-      Set<DependentChain> builder = new LinkedHashSet<>(allActiveDepChains.size());
-      // Retain only the ones which are applicable for this input batching decorator
-      for (DependentChain activeDepChain : allActiveDepChains) {
-        if (isApplicableToDependentChain.test(activeDepChain)) {
-          builder.add(activeDepChain);
-        }
-      }
-      this.activeDependantChains = unmodifiableSet(builder);
-      this.dependantChainsToFlush.addAll(this.activeDependantChains);
+    for (DepChainBatcherConfig depChainBatcherConfig : depChainBatcherConfigs) {
+      Set<DependentChain> dependentChains = depChainBatcherConfig.dependentChains();
+      dependentChainsToFlushByEpoch.add(
+          new LinkedHashSet<>(
+              // Retain only the ones which are applicable for this epoch
+              Sets.intersection(dependentChains, initiateActiveDepChains.dependentChains())));
     }
-  }
-
-  private Set<DependentChain> activeDependantChains() {
-    return activeDependantChains == null ? ImmutableSet.of() : activeDependantChains;
   }
 
   @SuppressWarnings({"UnnecessaryTypeArgument", "unchecked"}) // --> To Handle nullChecker errors
@@ -126,11 +143,5 @@ public final class InputBatchingDecorator
         }
       }
     }
-  }
-
-  @Override
-  public void onConfigUpdate(ConfigProvider configProvider) {
-    inputBatcher.onConfigUpdate(
-        new NestedConfig(String.format("input_batching.%s.", instanceId), configProvider));
   }
 }
