+++ krystex/src/main/java/com/flipkart/krystal/krystex/decoration/FlushCommand.java
@@ -2,4 +2,4 @@ package com.flipkart.krystal.krystex.decoration;
 
 import com.flipkart.krystal.krystex.kryon.DependentChain;
 
-public record FlushCommand(DependentChain dependantsChain) implements DecoratorCommand {}
+public record FlushCommand(DependentChain dependentChain) implements DecoratorCommand {}
