+++ krystex/src/main/java/com/flipkart/krystal/krystex/logicdecoration/OutputLogicDecorator.java
@@ -5,8 +5,13 @@ import com.flipkart.krystal.krystex.OutputLogicDefinition;
 import com.flipkart.krystal.krystex.kryon.KrystalExecutorCompletionListener;
 
 public non-sealed interface OutputLogicDecorator
-    extends LogicDecorator<OutputLogic<Object>, OutputLogicDefinition<Object>>,
-        KrystalExecutorCompletionListener {
+    extends LogicDecorator, KrystalExecutorCompletionListener {
 
-  OutputLogicDecorator NO_OP = (logicToDecorate, originalLogicDefinition) -> logicToDecorate;
+  OutputLogicDecorator NO_OP =
+      (logicToDecorate, originalLogicDefinition, context) -> logicToDecorate;
+
+  OutputLogic<Object> decorateLogic(
+      OutputLogic<Object> logicToDecorate,
+      OutputLogicDefinition<Object> originalLogicDefinition,
+      LogicExecutionContext context);
 }
