+++ krystex/src/main/java/com/flipkart/krystal/krystex/logicdecoration/OutputLogicDecoratorConfig.java
@@ -1,9 +1,9 @@
 package com.flipkart.krystal.krystex.logicdecoration;
 
 import com.flipkart.krystal.config.ConfigProvider;
-import com.flipkart.krystal.krystex.decoration.FlushCommand;
 import java.util.function.Function;
 import java.util.function.Predicate;
+import org.jspecify.annotations.Nullable;
 
 /**
  * Defines the configuration of an {@link OutputLogicDecorator}
@@ -18,23 +18,12 @@ import java.util.function.Predicate;
  *     {@link ConfigProvider} interface.
  * @param factory A factory which creates an instance of the logic decorator with the given
  *     instanceId.
- * @param enableFlushing true if the decorator needs to know when a dependant chain is {@link
- *     FlushCommand flushed}
  */
 public record OutputLogicDecoratorConfig(
     String decoratorType,
     Predicate<LogicExecutionContext> shouldDecorate,
     Function<LogicExecutionContext, String> instanceIdGenerator,
-    Function<OutputLogicDecoratorContext, OutputLogicDecorator> factory,
-    boolean enableFlushing) {
-
-  public OutputLogicDecoratorConfig(
-      String decoratorType,
-      Predicate<LogicExecutionContext> shouldDecorate,
-      Function<LogicExecutionContext, String> instanceIdGenerator,
-      Function<OutputLogicDecoratorContext, OutputLogicDecorator> factory) {
-    this(decoratorType, shouldDecorate, instanceIdGenerator, factory, false);
-  }
+    Function<OutputLogicDecoratorContext, @Nullable OutputLogicDecorator> factory) {
 
   public record OutputLogicDecoratorContext(
       String instanceId, LogicExecutionContext logicExecutionContext) {}
