+++ vajram/extensions/resilience4j/vajram-resilience4j/src/main/java/com/flipkart/krystal/vajram/resilience4j/bulkhead/Resilience4JBulkhead.java
@@ -52,7 +52,9 @@ public final class Resilience4JBulkhead implements OutputLogicDecorator {
 
   @Override
   public OutputLogic<Object> decorateLogic(
-      OutputLogic<Object> logicToDecorate, OutputLogicDefinition<Object> originalLogicDefinition) {
+      OutputLogic<Object> logicToDecorate,
+      OutputLogicDefinition<Object> originalLogicDefinition,
+      LogicExecutionContext context) {
     BulkheadAdapter bulkhead = this.adaptedBulkhead;
     if (bulkhead != null) {
       return input -> bulkhead.decorate(logicToDecorate, input);
