+++ vajram/extensions/resilience4j/vajram-resilience4j/src/main/java/com/flipkart/krystal/vajram/resilience4j/curcuitbreaker/Resilience4JCircuitBreaker.java
@@ -39,7 +39,9 @@ public final class Resilience4JCircuitBreaker implements OutputLogicDecorator {
 
   @Override
   public OutputLogic<Object> decorateLogic(
-      OutputLogic<Object> logicToDecorate, OutputLogicDefinition<Object> originalLogicDefinition) {
+      OutputLogic<Object> logicToDecorate,
+      OutputLogicDefinition<Object> originalLogicDefinition,
+      LogicExecutionContext context) {
     CircuitBreaker circuitBreaker = this.circuitBreaker;
     if (circuitBreaker != null) {
       return input ->
